backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lebrq-backend"
version = "0.1.0"
description = "FastAPI backend for LebrqApp"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.2.1",
  "SQLAlchemy>=2.0.25",
  "aiomysql>=0.2.0",
  "asyncmy>=0.2.9",
  "alembic>=1.13.1",
  "python-dotenv>=1.0.1",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.8.0",
  "httpx>=0.27.2",
  "orjson>=3.10.7",
  "uvloop; platform_system != 'Windows'",
  "python-multipart>=0.0.9",
  "aiosqlite>=0.19.0",
]

[tool.uvicorn]
reload = true
factory = true
